Refactor: store 디렉토리 src 내부로 옮김. store => src/store
