+++ src/components/Send/VoiceModal/VoiceModal.tsx
@@ -1,6 +1,6 @@
 import styles from "./VoiceModal.module.scss";
 import { useReactMediaRecorder } from "react-media-recorder";
-import { useLetterFormStore } from "../../../../store/useLetterFormStore";
+import { useLetterFormStore } from "../../../store/useLetterFormStore";
 
 const VoiceModal = () => {
   const audio = useLetterFormStore((state) => state.audio);
